Create README.md
Organising work in the “new normal” brings its own set of challenges and asks us to learn new skills and adjust our tools. In short, it requires us to become incredibly flexible when it comes to managing tasks, teammates, and work in general.
This Task tracker is a tool that helps you meet task deadlines and progress swiftly in the workplace.